pyproject.toml
[project]
name = "toolsbot"
version = "0.1.0"
description = "工具机器人"
authors = [
    {name = "zhangzhi", email = "781029588@qq.com"},
]
dependencies = [
    "nonebot2>=2.2.1",
    "fastapi>=0.110.1",
    "httpx>=0.27.0",
    "tortoise-orm[asyncpg]>=0.20.0",
    "aerich>=0.7.2",
    "python-jose[ecdsa]>=3.3.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "nb-cli>=1.4.1",
]

[tool.nonebot]
adapters = [
    
]
plugins = []
plugin_dirs = ["toolsbot/plugins"]
builtin_plugins = []

[tool.pyright]
include = ["toolsbot"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
]
# reportIncompatibleMethodOverride = "none"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./",
  ] },
  { root = "./" },
]

typeCheckingMode = "basic"
reportShadowedImports = false
disableBytesTypePromotions = true
